proyecto_mintic/requirements.txt
# ── Extracción ────────────────────────────────────────────────────
requests==2.31.0
selenium==4.18.1
scrapy==2.11.0
python-dotenv==1.0.1

# ── Procesamiento ─────────────────────────────────────────────────
pandas==2.2.0
openpyxl==3.1.2
pdfplumber==0.10.3

# ── Modelo IA ─────────────────────────────────────────────────────
prophet>=1.1.5
scikit-learn>=1.4.0
statsmodels>=0.14.0          # fallback si CmdStan no compila en Windows

# ── Base de datos ─────────────────────────────────────────────────
psycopg2-binary==2.9.9
sqlalchemy==2.0.28
